requirements.txt
streamlit
PyPDF2
camelot-py[cv]
pytesseract
pillow
sentence-transformers
torch
clip
chromadb
requests
